Back web

+++ site-performee/src/models/componenteModel.js
@@ -1,31 +1,47 @@
 var database = require("../database/config")
 
 function selecionarTudo() {
-    var instrucao = `
-        SELECT * FROM Componente;
-    `;
+    if (process.env.AMBIENTE_PROCESSO == "produção") {
+
+        // script sqlServer
+
+    } else if (process.env.AMBIENTE_PROCESSO == "desenvolvimento") {
+        var instrucao = `
+            SELECT * FROM Componente;
+        `;
+    } else {
+        console.log('Ambienetes não definidos no app.js');
+        return;
+    }
     return database.executar(instrucao);
 }
 
 function deletarComponente(tipo, id) {
-    if (tipo == 'DC') {
-        var instrucao = `
-        delete from Componente where fkDataCenter = '${id}';
+    if (process.env.AMBIENTE_PROCESSO == "produção") {
+
+        // script sqlServer
+
+    } else if (process.env.AMBIENTE_PROCESSO == "desenvolvimento") {
+        if (tipo == 'DC') {
+            var instrucao = `
+            delete from Componente where fkDataCenter = '${id}';
+            `;
+        } else if (tipo == 'Server') {
+            var instrucao = `
+        delete from Componente where fkServidor = '${id}';
         `;
-        return database.executar(instrucao);
-    } else if (tipo == 'Server') {
-        var instrucao = `
-    delete from Componente where fkServidor = '${id}';
-    `;
-    return database.executar(instrucao);
+        }
+        else {
+            var instrucao = `
+        delete from Componente where fkEmpresa = '${id}';
+        `;
+        }
+    } else {
+        console.log('Ambienetes não definidos no app.js');
+        return;
     }
-    else {
-        var instrucao = `
-    delete from Componente where fkEmpresa = '${id}';
-    `;
     return database.executar(instrucao);
-    }
-  }
+}
 
 module.exports = {
     selecionarTudo,
